--- homepage_options.json
@@ -1,1 +1,0 @@
-{"cellOptions": ["dining", "laundry", "studyRoomBooking"]}
